requirements.txt

# no versioning for now
# conda create --name viz-challenge-2024 python==3.10

pudb
dash
pandas
attrdictionary
yml 
pyyaml
